requirements.txt
lightning
matplotlib
numpy
pandas
pyperclip
PyYAML
selenium
tensorboard
timm
## package location
# --find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchaudio
torchmetrics
torchvision
webdriver_manager
